Plugs the Decimal assertions into the Spike code.
Comments the useless IntFluentAssertion and StringFluentAssertion, since the factory of the Spike.Check class automatically instantiates FluentAssertion<T> wrappers.

Keeps the factory mechanism since it will be useful to provide a way to inject FluentAssertion specific types with more than a Sut as property (for web response and headers for instance).